// types/supabase.ts
export type Database = {
  public: {
    Tables: {
// ...
      assistants: {
        Row: {
          created_at: string
          description: string | null
          id: string
          name: string
          openai_assistant_id: string | null
          openai_vector_store_id: string | null
          system_prompt: string | null
          user_prompt: string | null
          welcome_message: string[] | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: string
          name: string
          openai_assistant_id?: string | null
          openai_vector_store_id?: string | null
          system_prompt?: string | null
          user_prompt?: string | null
          welcome_message?: string[] | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: string
          name?: string
          openai_assistant_id?: string | null
          openai_vector_store_id?: string | null
          system_prompt?: string | null
          user_prompt?: string | null
          welcome_message?: string[] | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      increment_job_processed_emails: {
        Args: { p_job_id: string; p_increment_amount?: number }
        Returns: undefined
      }
    }
    Enums: {
      email_priority: "LOW" | "MID" | "HIGH"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}